Created the books app (#48)
* Updated the version being worked on

* Moved the Books model to the books app

* Added the books app to installed appps and ran DB migrations

* Registered the Books model to the books app admin

* Updated URLconf for books

* Updated views for books

* Updated functional tests

* Updated docs

* Updated resources migration 0002 because of an SQLite issue
