--- docs/accounts.tests.rst
@@ -1,37 +1,0 @@
-accounts.tests package
-======================
-
-Submodules
-----------
-
-accounts.tests.test\_models module
-----------------------------------
-
-.. automodule:: accounts.tests.test_models
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-accounts.tests.test\_urls module
---------------------------------
-
-.. automodule:: accounts.tests.test_urls
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-accounts.tests.test\_views module
----------------------------------
-
-.. automodule:: accounts.tests.test_views
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: accounts.tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
