+++ docs/books.rst
@@ -1,0 +1,53 @@
+books package
+=============
+
+Submodules
+----------
+
+books.admin module
+------------------
+
+.. automodule:: books.admin
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+books.apps module
+-----------------
+
+.. automodule:: books.apps
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+books.models module
+-------------------
+
+.. automodule:: books.models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+books.urls module
+-----------------
+
+.. automodule:: books.urls
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+books.views module
+------------------
+
+.. automodule:: books.views
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: books
+   :members:
+   :undoc-members:
+   :show-inheritance:
