--- docs/resources.tests.rst
@@ -1,37 +1,0 @@
-resources.tests package
-=======================
-
-Submodules
-----------
-
-resources.tests.test\_models module
------------------------------------
-
-.. automodule:: resources.tests.test_models
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-resources.tests.test\_urls module
----------------------------------
-
-.. automodule:: resources.tests.test_urls
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-resources.tests.test\_views module
-----------------------------------
-
-.. automodule:: resources.tests.test_views
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: resources.tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
