+++ resources/tests/test_urls.py
@@ -15,14 +15,6 @@ class ResourcesURLsTestCase(TestCase):
         root = resolve('/')
         self.assertEqual(root.func, index)
 
-    def test_book_list_url(self):
-        """
-        Test that the URL for book list resolves to the
-        correct view function
-        """
-        book_list = resolve(reverse('books'))
-        self.assertEqual(book_list.func.__name__, 'BookListView')
-
     def test_video_list_url(self):
         """
         Test that the URL for video list resolves to the
@@ -30,16 +22,6 @@ class ResourcesURLsTestCase(TestCase):
         """
         video_list = resolve(reverse('videos'))
         self.assertEqual(video_list.func.__name__, 'VideoListView')
-
-    def test_book_details_url(self):
-        """
-        Test that the URL for book details resolves to the
-        correct view function
-        """
-        book_details = resolve(reverse('book',
-            kwargs={'slug': 'the-hydroponics-handbook'}))
-        self.assertEqual(book_details.func.__name__,
-            'BookDetailView')
 
     def test_video_details_url(self):
         """
